+++ CHANGELOG.md
@@ -25,7 +25,7 @@
   Will use the .no-molly-guard version of reboot command if exists.
   Use ayatana app indicator if available.
 
-  * 0.0.11 (2026-10-04): Icon is now bundled with the package
+  * 0.0.11 (2025-10-04): Icon is now bundled with the package
   This also allows it to work on Linux Mint.
   Uses Ayatana app indicator if available.
   grub-reboot-picker now runs as normal user instead of root.